+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Version 1.0.x
+
+### 1.0.0.x (2024-Oct-25)
+
+- Updated example to use with BACnetStack version 5.1.2.0
+- Added an example of Out Of Service
+
 ## Version 0.0.x
 
 ### 0.0.26.x (2024-May-02)